requirements.txt

ipython
jupyter
python-dotenv
numpy
scipy
sklearn
cached-property
tqdm
tensorflow
keras
gensim
attrs
editdistance
wordfreq
boltons
textblob
matplotlib
click
rpy2
ujson
pandas
annoy
torchtext
